+++ Utilities/otbossim/src/ossim/imaging/ossimRgbGridRemapEngine.cpp
@@ -14,7 +14,7 @@
 // LIMITATIONS: None.
 //
 //*****************************************************************************
-//  $Id: ossimRgbGridRemapEngine.cpp 9094 2006-06-13 19:12:40Z dburken $
+//  $Id: ossimRgbGridRemapEngine.cpp 15833 2009-10-29 01:41:53Z eshirschorn $
 
 #include <ossim/imaging/ossimRgbGridRemapEngine.h>
 
@@ -166,7 +166,7 @@ void ossimRgbGridRemapEngine::assignRemapValues (
    // Declare a 2D array that will contain all of the contributing sources'
    // RGB mean values. Also declare the accumulator target vector.
    //***
-   int num_contributors =  sources_list.size();
+   int num_contributors =  (int)sources_list.size();
    double** contributor_pixel = new double* [num_contributors];
    for (i=0; i<num_contributors; i++)
       contributor_pixel[i] = new double[3];
